# feat: Zero-Copy Lazy JSON Parser with SIMD Acceleration (v0.2.1)