--- a/crates/pjs-core/src/application/services/session_service.rs
+++ b/crates/pjs-core/src/application/services/session_service.rs
@@ -1,3 +1,4 @@
+
 //! High-level session management service
 
 use crate::{